requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
--extra-index-url https://www.piwheels.org/simple

aiosqlite==0.16.1
    # via
    #   -r requirements.in
    #   tortoise-orm
attrs==20.3.0
    # via
    #   automat
    #   twisted
automat==20.2.0
    # via twisted
bleak==0.10.0
    # via -r requirements.in
click==7.1.2
    # via uvicorn
constantly==15.1.0
    # via twisted
fastapi==0.63.0
    # via -r requirements.in
gunicorn==20.0.4
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
httptools==0.1.1
    # via uvicorn
hyperlink==21.0.0
    # via twisted
idna==3.1
    # via hyperlink
incremental==17.5.0
    # via twisted
iso8601==0.1.13
    # via tortoise-orm
nodeenv==1.5.0
    # via -r requirements.in
pydantic==1.7.3
    # via fastapi
pyhamcrest==2.0.2
    # via twisted
pypika==0.44.1
    # via tortoise-orm
python-dotenv==0.15.0
    # via uvicorn
pytz==2020.5
    # via tortoise-orm
pyyaml==5.4.1
    # via uvicorn
six==1.15.0
    # via
    #   automat
    #   txdbus
starlette==0.13.6
    # via fastapi
tortoise-orm==0.16.20
    # via -r requirements.in
twisted==20.3.0
    # via txdbus
txdbus==1.1.2
    # via bleak
typing-extensions==3.7.4.3
    # via aiosqlite
uvicorn[standard]==0.13.3
    # via -r requirements.in
uvloop==0.14.0
    # via uvicorn
watchgod==0.6
    # via uvicorn
websockets==8.1
    # via uvicorn
zope.interface==5.2.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools